+++ src/components/swap/FeesAndRateDetails.tsx
@@ -21,11 +21,13 @@ const RateDisplay = ({
   inputAsset,
   outputAsset,
   formattedRate,
+  formattedTokenPrice,
   className = "",
 }: {
   inputAsset?: Asset;
   outputAsset?: Asset;
-  formattedRate: number;
+  formattedRate?: number;
+  formattedTokenPrice?: number;
   className?: string;
 }) => (
   <div className={`flex min-w-fit items-center gap-1`}>
@@ -41,7 +43,8 @@ const RateDisplay = ({
       weight="regular"
       className={`!text-nowrap ${className}`}
     >
-      {formattedRate} {outputAsset?.symbol}
+      {formattedRate && `${formattedRate} ${outputAsset?.symbol}`}
+      {formattedTokenPrice && `$${formattedTokenPrice}`}
     </Typography>
   </div>
 );
@@ -51,8 +54,14 @@ export const FeesAndRateDetails = () => {
   const [isHovered, setIsHovered] = useState(false);
   const targetRef = useRef<HTMLDivElement>(null);
 
-  const { inputAsset, outputAsset, rate, networkFees, showComparisonHandler } =
-    swapStore();
+  const {
+    inputAsset,
+    outputAsset,
+    rate,
+    networkFees,
+    showComparisonHandler,
+    fiatTokenPrices,
+  } = swapStore();
   const { account: btcAddress } = useBitcoinWallet();
   const { solanaAddress } = useSolanaWallet();
   const { address } = useEVMWallet();
@@ -61,6 +70,11 @@ export const FeesAndRateDetails = () => {
   const formattedRate = useMemo(
     () => formatAmount(rate, 0, isBitcoinChains ? 7 : 3),
     [isBitcoinChains, rate]
+  );
+
+  const formattedTokenPrice = useMemo(
+    () => formatAmount(fiatTokenPrices.input, 0, 2),
+    [fiatTokenPrices.input]
   );
 
   const refundAddress = useMemo(
@@ -134,7 +148,7 @@ export const FeesAndRateDetails = () => {
                 <RateDisplay
                   inputAsset={inputAsset}
                   outputAsset={outputAsset}
-                  formattedRate={formattedRate}
+                  formattedTokenPrice={formattedTokenPrice}
                   className="!text-mid-grey"
                 />
               </motion.div>
@@ -152,7 +166,7 @@ export const FeesAndRateDetails = () => {
                 <RateDisplay
                   inputAsset={inputAsset}
                   outputAsset={outputAsset}
-                  formattedRate={formattedRate}
+                  formattedTokenPrice={formattedTokenPrice}
                 />
               </motion.div>
             ) : (
@@ -165,7 +179,7 @@ export const FeesAndRateDetails = () => {
                   <GasStationIcon className="h-3 w-3" />
                   <Typography
                     size="h5"
-                    weight="medium"
+                    weight="regular"
                     className="!text-nowrap"
                   >
                     {networkFees === 0 ? (