+++ src/hooks/useSwap.tsx
@@ -51,6 +51,7 @@ export const useSwap = () => {
     swapAssets,
     setAsset,
     setIsFetchingQuote,
+    setFiatTokenPrices,
     isComparisonVisible,
     setIsValidBitcoinAddress,
     // setIsApproving,
@@ -236,7 +237,7 @@ export const useSwap = () => {
           );
           // Add network fee to output amount before calculating rate
           let outputAmountWithFee = Number(quoteAmountInDecimals);
-          if (!isBitcoin(fromAsset.chain) && !isBitcoin(toAsset.chain)) {
+          if (fromAsset.symbol === "USDC" && toAsset.symbol === "USDC") {
             outputAmountWithFee = Number(quoteAmountInDecimals) + networkFees;
           }
           const rate = outputAmountWithFee / Number(amount);
@@ -262,7 +263,10 @@ export const useSwap = () => {
           const outputTokenPrice = outputAmount.multipliedBy(
             quote.val.output_token_price
           );
-
+          setFiatTokenPrices({
+            input: quote.val.input_token_price.toString(),
+            output: quote.val.output_token_price.toString(),
+          });
           setTokenPrices({
             input: inputTokenPrice.toString(),
             output: outputTokenPrice.toString(),