+++ src/store/assetInfoStore.ts
@@ -39,6 +39,8 @@ import { Hex } from "viem";
 import { getSpendableBalance } from "../utils/getmaxBtc";
 import logger from "../utils/logger";
 import { getLegacyGasEstimate } from "../utils/getNativeTokenFee";
+import { SupportedChains } from "../layout/wagmi/config";
+import { getAllWorkingRPCs } from "../utils/rpcUtils";
 
 export type Networks = {
   [chain in Chain]: ChainData & { assetConfig: Omit<AssetConfig, "chain">[] };
@@ -74,6 +76,7 @@ type AssetInfoState = {
   chains: Chains | null;
   fiatData: Record<string, number | undefined>;
   balances: Record<string, BigNumber | undefined>;
+  workingRPCs: Record<number, string[]>;
   isLoading: boolean;
   isAssetSelectorOpen: {
     isOpen: boolean;
@@ -87,6 +90,7 @@ type AssetInfoState = {
   };
   setOpenAssetSelector: (type: IOType) => void;
   CloseAssetSelector: () => void;
+  fetchAndSetRPCs: () => Promise<void>;
   fetchAndSetAssetsAndChains: () => Promise<void>;
   fetchAndSetStrategies: () => Promise<void>;
   fetchAndSetFiatValues: () => Promise<void>;
@@ -111,6 +115,7 @@ export const assetInfoStore = create<AssetInfoState>((set, get) => ({
   allChains: null,
   fiatData: {},
   balances: {},
+  workingRPCs: {},
   isAssetSelectorOpen: {
     isOpen: false,
     type: IOType.input,
@@ -122,7 +127,11 @@ export const assetInfoStore = create<AssetInfoState>((set, get) => ({
     error: null,
     isLoading: false,
   },
-
+  fetchAndSetRPCs: async () => {
+    set({ isLoading: true });
+    const workingRPCs = await getAllWorkingRPCs([...SupportedChains]);
+    set({ workingRPCs, isLoading: false });
+  },
   setOpenAssetSelector: (type) =>
     set({
       isAssetSelectorOpen: {
@@ -236,7 +245,7 @@ export const assetInfoStore = create<AssetInfoState>((set, get) => ({
     address: string,
     fetchOnlyAsset?: Asset
   ) => {
-    const { assets } = get();
+    const { assets, workingRPCs } = get();
     if (!assets) return;
 
     let tokensByChain: Partial<Record<Chain, Asset[]>> = {};
@@ -256,7 +265,8 @@ export const assetInfoStore = create<AssetInfoState>((set, get) => ({
           const chainBalances = await getBalanceMulticall(
             assets.map((asset) => asset.tokenAddress) as Hex[],
             address as Hex,
-            chain as EvmChain
+            chain as EvmChain,
+            workingRPCs
           );
 
           const updatedBalances: Record<string, BigNumber | undefined> = {};