+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.1.5]
+
+### Fixed
+
+- Updated balance retrieval function to use fallback RPC clients as default instead of default wallet RPCs.
+- Rate conversion and display estimated token price for inputAsset.
+- Fixed fiat token prices.
+
+
 ## [1.1.4]
 
 ### Fixed